Untested Fix
Having NATTEN compile issues but want to address issue 16. Will fix
NATTEN compile and then merge.

Fixed some hydra resolvers but also added a default conf which should
make this former pointless.
