+++ requirements.txt
@@ -1,15 +1,19 @@
+torch
+torchvision
 lmdb
 timm
 scipy
-sklearn
+scikit-learn
 einops
 tqdm
 wandb
 ninja
 natten>=0.14.6
 hydra-core
+hydra_colorlog
 joblib
 dill
 imageio-ffmpeg
 ftfy
 matplotlib
+rich
